chore: move get-search-results to /clients/queries
