+++ core/client/queries/get-search-results.ts
